# perf(ext/node): optimize http headers